fixed types in storage. small fixes in explorer

+++ core/storage/migrations/2019-04-02-100645_proofs/up.sql
@@ -1,14 +1,14 @@
 -- Your SQL goes here
 
 CREATE TABLE proofs (
-    block_number    serial primary key,
+    block_number    bigserial primary key,
     proof           jsonb not null,
     created_at      timestamp not null default now()
 );
 
 CREATE TABLE prover_runs (
     id              serial primary key,
-    block_number    int not null,
+    block_number    bigint not null,
     worker          text,
     created_at      timestamp not null default now(),
     updated_at      timestamp not null default now()
